+++ src/main/java/kr/co/springbootweeklymission/member/application/MemberService.java
@@ -2,10 +2,10 @@ package kr.co.springbootweeklymission.member.application;
 
 import kr.co.springbootweeklymission.common.error.exception.NotFoundException;
 import kr.co.springbootweeklymission.common.response.ResponseStatus;
-import kr.co.springbootweeklymission.member.api.dto.request.MemberReqDTO;
-import kr.co.springbootweeklymission.member.api.dto.response.MemberResDTO;
 import kr.co.springbootweeklymission.member.domain.entity.Member;
 import kr.co.springbootweeklymission.member.domain.repository.MemberRepository;
+import kr.co.springbootweeklymission.member.presentation.dto.request.MemberReqDTO;
+import kr.co.springbootweeklymission.member.presentation.dto.response.MemberResDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
