+++ src/main/java/kr/co/springbootweeklymission/member/domain/entity/Member.java
@@ -1,7 +1,7 @@
 package kr.co.springbootweeklymission.member.domain.entity;
 
-import kr.co.springbootweeklymission.member.api.dto.request.MemberReqDTO;
 import kr.co.springbootweeklymission.member.domain.model.MemberStatus;
+import kr.co.springbootweeklymission.member.presentation.dto.request.MemberReqDTO;
 import lombok.*;
 
 import java.util.UUID;
