+++ src/main/java/kr/co/springbootweeklymission/voucher/domain/entity/Voucher.java
@@ -2,8 +2,8 @@ package kr.co.springbootweeklymission.voucher.domain.entity;
 
 import kr.co.springbootweeklymission.common.error.exception.WrongVoucherPolicyException;
 import kr.co.springbootweeklymission.common.response.ResponseStatus;
-import kr.co.springbootweeklymission.voucher.api.dto.request.VoucherReqDTO;
 import kr.co.springbootweeklymission.voucher.domain.model.VoucherPolicy;
+import kr.co.springbootweeklymission.voucher.presentation.dto.request.VoucherReqDTO;
 import lombok.*;
 
 import java.util.UUID;
