+++ src/main/java/kr/co/springbootweeklymission/voucher/domain/repository/InFileVoucherRepository.java
@@ -4,8 +4,8 @@ import kr.co.springbootweeklymission.common.error.exception.FileIOException;
 import kr.co.springbootweeklymission.common.error.exception.NotSupportedException;
 import kr.co.springbootweeklymission.common.response.ResponseStatus;
 import kr.co.springbootweeklymission.common.util.FileConverter;
-import kr.co.springbootweeklymission.voucher.api.dto.response.VoucherResDTO;
 import kr.co.springbootweeklymission.voucher.domain.entity.Voucher;
+import kr.co.springbootweeklymission.voucher.presentation.dto.response.VoucherResDTO;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
