+++ src/main/java/kr/co/springbootweeklymission/wallet/presentation/dto/request/WalletReqDTO.java
@@ -1,4 +1,4 @@
-package kr.co.springbootweeklymission.wallet.api.dto.request;
+package kr.co.springbootweeklymission.wallet.presentation.dto.request;
 
 import lombok.*;
