+++ src/test/java/kr/co/springbootweeklymission/voucher/creators/VoucherCreators.java
@@ -1,8 +1,8 @@
 package kr.co.springbootweeklymission.voucher.creators;
 
-import kr.co.springbootweeklymission.voucher.api.dto.request.VoucherReqDTO;
 import kr.co.springbootweeklymission.voucher.domain.entity.Voucher;
 import kr.co.springbootweeklymission.voucher.domain.model.VoucherPolicy;
+import kr.co.springbootweeklymission.voucher.presentation.dto.request.VoucherReqDTO;
 
 import java.util.UUID;
